// frontend/src/api-generated/models/danhMucCauHinhRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface DanhMucCauHinhRequest {
  maCauHinh?: string;
  tenCauHinh: string;
  moTaCauHinh?: string;
  nhomCauHinh?: string;
  loaiDuLieu?: string;
  giaTriMacDinh?: string;
  trangThai?: string;
}